Remove v1 tests
